requirements.txt
pytest~=8.4.1
selenium~=4.33.0
allure-pytest
pytest-html
allure-python-commons~=2.14.3